# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: FreeKick Gacha Card Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market:
    get:
      operationId: getMarket
      tags: [market]
      summary: List active market auctions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: position
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Market listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketPage"
...
  /market/list:
    post:
      operationId: createListing
      tags: [market]
      summary: List a card for auction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingInput"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListing"
        "422":
          description: Bid below minimum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /market/bid/{listingId}:
    post:
      operationId: placeBid
      tags: [market]
      summary: Place a bid on a market listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBidInput"
      responses:
        "200":
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListing"
        "422":
          description: Bid too low or insufficient coins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /market/cancel/{listingId}:
    delete:
      operationId: cancelListing
      tags: [market]
      summary: Cancel a listing (only if no bids)
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        min_bid:
          type: integer
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    MarketListing:
      type: object
      required: [id, seller_id, seller_username, inventory_id, player_id, player_name, player_initials, player_overall, player_position, player_club, player_country, player_is_gk, starting_bid, current_bid, expires_at, created_at, quick_sell_value]
      properties:
        id:
          type: integer
        seller_id:
          type: integer
        seller_username:
          type: string
        inventory_id:
          type: integer
        player_id:
          type: integer
        player_name:
          type: string
        player_initials:
          type: string
        player_overall:
          type: integer
        player_position:
          type: string
        player_club:
          type: string
        player_country:
          type: string
        player_is_gk:
          type: integer
        starting_bid:
          type: integer
        current_bid:
          type: integer
        current_bidder_id:
          type: ["integer", "null"]
        current_bidder_username:
          type: ["string", "null"]
        expires_at:
          type: integer
        created_at:
          type: string
        quick_sell_value:
          type: integer
    MarketPage:
      type: object
      required: [listings, page, total, pages]
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/MarketListing"
        page:
          type: integer
        total:
          type: integer
        pages:
          type: integer
    CreateListingInput:
      type: object
      required: [inventory_id, starting_bid, duration]
      properties:
        inventory_id:
          type: integer
        starting_bid:
          type: integer
        duration:
          type: string
          enum: ["1h", "2h", "6h", "12h", "24h"]
    PlaceBidInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
...