# lib/api-zod/src/generated/types/marketListing.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FreeKick Gacha Card Game API
 * OpenAPI spec version: 0.2.0
 */

export interface MarketListing {
  id: number;
  seller_id: number;
  seller_username: string;
  inventory_id: number;
  player_id: number;
  player_name: string;
  player_initials: string;
  player_overall: number;
  player_position: string;
  player_club: string;
  player_country: string;
  player_is_gk: number;
  starting_bid: number;
  current_bid: number;
  /** @nullable */
  current_bidder_id?: number | null;
  /** @nullable */
  current_bidder_username?: string | null;
  expires_at: number;
  created_at: string;
  quick_sell_value: number;
}